Move big sensor list into its own file
